\name{EMLab-package}
\alias{EMLab-package}
\alias{EMLab}
\docType{package}
\title{
\packageTitle{EMLab}
}
\description{
\packageDescription{EMLab}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{EMLab}

Maintainer: \packageMaintainer{EMLab}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = EMLab)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{EMLab}}
%\section{Documentation Index}{\packageIndices{EMLab}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
